requirements.txt
streamlit
google-generativeai
faiss-cpu
sentence-transformers
speechrecognition
pyttsx3
pyaudio
Pillow
